bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
fastapi>=0.100
uvicorn>=0.22
openpyxl>=3.1.2
